Change from std::thread to std::future.
I changed the use of the threads to be done through the wrap object std::future, this is because it abstracts and handles the lifetime of threads easier. I also forgot to add the printing of the top 10 prime numbers, so that was done. Lastly, I added an empty input prompt, since when the program runs through compiled binaries, it would automatically close once it finished.